+++ wims/src/Misc/scienceprint/scienceprint.c
@@ -1,11 +1,11 @@
 /*
 
 *********************************************************************************
-* J.M. Evers 3/2012								*
-* This is all amateur scriblings... So no copyrights.				*
-* This source code file, and compiled objects derived from it,			*
-* can be used and distributed without restriction, including for commercial use	*
-* No warrenty whatsoever							*
+* J.M. Evers 3/2012                                                             *
+* This is all amateur scriblings... So no copyrights.                           *
+* This source code file, and compiled objects derived from it,                  *
+* can be used and distributed without restriction, including for commercial use *
+* No warrenty whatsoever                                                        *
 *********************************************************************************
 20/6/2012
 Corrected significance flaw when using prefixes
@@ -43,34 +43,34 @@ type = 0 : calc   notation : 120000,3,0 -> 1.20*10^5
 type = 1 : html notation   : 120000,3,1 -> 1.20&times;10<sup>5</sup>
 type = 2 : latex notation  : 120000,3,2 -> 1.20 \times 10^{5}
 type = 3 : prefix-notation : 120000,3,3 -> 120.0 k
-type = 3 : if -24 > prefix > 24  	-> type = 1 (html)
+type = 3 : if -24 > prefix > 24         -> type = 1 (html)
 type = 4 : mathml notation 
 
 multiple conversion: use space between arguments
 scienceprint 120000,4 122900,5 120036,6,3 --> 120.0*10^3,122.90*10^3,120.036 k
 
-24  yotta 	Y 
-21  zetta 	Z 
-18  exa 	E 
-15  peta	P 
-12  tera 	T
-9   giga 	G
-6   mega 	M
-3   kilo 	k 
+24  yotta       Y 
+21  zetta       Z 
+18  exa         E 
+15  peta        P 
+12  tera        T
+9   giga        G
+6   mega        M
+3   kilo        k 
 
-2   hecto	h
-1   deca 	da
--1  deci 	d 
--2  centi 	c 
+2   hecto       h
+1   deca        da
+-1  deci        d 
+-2  centi       c 
 
--3  milli 	m
--6  micro 	µ
--9  nano 	n
--12 pico 	p
--15 femto 	f
--18 atto 	a
--21 zepto 	z
--24 yocto 	y
+-3  milli       m
+-6  micro       µ
+-9  nano        n
+-12 pico        p
+-15 femto       f
+-18 atto        a
+-21 zepto       z
+-24 yocto       y
 
 */
 
@@ -95,7 +95,7 @@ char *str_replace ( const char *word, const char *sub_word, const  char *rep_wor
     new_word = strdup (word);
     head = new_word;
     while ( (part_word = strstr ( head, sub_word ))){
-	old_word = new_word;
+        old_word = new_word;
         new_word = malloc ( strlen ( old_word ) - strlen ( sub_word ) + strlen ( rep_word ) + 1 );
         /*failed to alloc mem, free old word and return NULL */
         if ( new_word == NULL ){
@@ -117,7 +117,7 @@ char *printscience(double value, int sig, int format , int cnt ,int size){
     static char *plus[] = {"","k", "M", "G", "T", "P", "E", "Z", "Y" };
     static char *min_word[] = {"","milli","micro","nano","pico","femto","atto","zepto","yocto"};
     static char *plus_word[] = {"","kilo", "mega", "giga", "tera", "peta", "exa", "zetta", "yotta" };
-    char *sign = NULL;char *prefix = NULL;int pm;int factor;
+    char *sign = NULL;char *prefix = NULL;double pm;int factor;
     int exponent10 = 0;
     int use_word = 0;if(format == 5){format = 3; use_word = 1;} /* switch to using words in stead of prefix  */
     if(value < 0.0) {pm = -0.5; sign = "-";value = -value;} else {sign = ""; pm = 0.5;}    
@@ -126,32 +126,32 @@ char *printscience(double value, int sig, int format , int cnt ,int size){
      no significance truncations...just science notation 1234 -> 1.234*10^3 
      try (!) to use same amount of digits
      */
-	sig = size;
-	if(format == 3){format = 1;} /* never prefix --> html notation */
+        sig = size;
+        if(format == 3){format = 1;} /* never prefix --> html notation */
     }
     if(value == 0){fprintf(stdout, "%s%.*f", sign, sig-1, value);return NULL;} /* no need to go further */
     if(value>1){
-	while(value >= 10){
-	    value=value / 10.0;
-	    exponent10++;
-	    /* need to set a limit to number of while loops ! */
-	    if(exponent10 > 100){fprintf(stdout,"error : number too big (exponent > 100)\n");return 0;}
-	}
+        while(value >= 10){
+            value=value / 10.0;
+            exponent10++;
+            /* need to set a limit to number of while loops ! */
+            if(exponent10 > 100){fprintf(stdout,"error : number too big (exponent > 100)\n");return 0;}
+        }
     }
     else /* 0 < value < 1 --> exponent10 < 0 */
     {
-	while(value < 1){
-	    value=value*10;
-	    exponent10--;
-	    /* need to set a limit to number of while loops ! */
-	    if(exponent10 <-100){fprintf(stdout,"error : number too small (exponent < -100)\n");return 0;}
-	}
+        while(value < 1){
+            value=value*10;
+            exponent10--;
+            /* need to set a limit to number of while loops ! */
+            if(exponent10 <-100){fprintf(stdout,"error : number too small (exponent < -100)\n");return 0;}
+        }
     }
     /* 27/9/2013 avoid truncating and do rounding...but not with 7 significant digits*/
     if(sig > 6){factor = 100000;}else{factor = pow(10,sig+1);}
     value = (round(factor*value + (pm) ))/factor; /* pm = +/- 0.5 */
     if(format == 3 && ((exponent10 < PREFIX_START) || (exponent10 > PREFIX_END))){
-	format = 1; /* not in my list of prefixes ; print in html ! */
+        format = 1; /* not in my list of prefixes ; print in html ! */
     }
     sig = sig - 1; /* "%.*f" counts the "." */
     if(cnt > 1){fprintf(stdout,",");}/* more than one conversion to do : make list */
@@ -159,10 +159,10 @@ char *printscience(double value, int sig, int format , int cnt ,int size){
     if(exponent10 == 0){format = 6;} /* no need for 2*10^0 */
     if(sig < 0){sig = 0;} /* better be safe than sorry... */
     switch(format){
-	case 0: fprintf(stdout, "%s%.*f*10^%d", sign, sig, value, exponent10);break;
-	case 1:	fprintf(stdout, "%s%.*f&times;10<sup>%d</sup>", sign, sig, value, exponent10);break;
-	case 2: fprintf(stdout, "%s%.*f \\times 10^{%d}", sign, sig, value, exponent10);break;
-	case 3:
+        case 0: fprintf(stdout, "%s%.*f*10^%d", sign, sig, value, exponent10);break;
+        case 1: fprintf(stdout, "%s%.*f&times;10<sup>%d</sup>", sign, sig, value, exponent10);break;
+        case 2: fprintf(stdout, "%s%.*f \\times 10^{%d}", sign, sig, value, exponent10);break;
+        case 3:
 /*
 1,1,3 -> 1
 10,1,3 -> 1*10^-2 k
@@ -185,27 +185,27 @@ char *printscience(double value, int sig, int format , int cnt ,int size){
 0.00000001,1,3-> 1*10^-2 µ
 0.000000001,1,3-> 1 n
 */
-	exp = exponent10%3;
-	idx = round(exponent10/3);
-	if( exponent10 > 0  ){
-	    if(use_word == 0 ){ prefix = plus[idx]; } else { prefix = plus_word[idx]; }
-	}
-	else
-	{
-	    if(use_word == 0){ prefix = min[-1*idx]; } else { prefix = min_word[-1*idx]; }
-	}
-	if( exp == 0){
-	    fprintf(stdout, "%s%.*f %s",sign, sig, value,prefix);
-	}
-	else
-	{
-	    fprintf(stdout, "%s%.*f&times;10<sup>%d</sup> %s", sign, sig, value, exp, prefix);
-	}
-	break;
-	case 4: fprintf(stdout, "<math xmlns=\"http://www.w3.org/1998/Math/MathML\" display=\"inline\"><mstyle id=\"wims_mathml\" mathsize=\"110%%\"><mn>%s%.*f</mn><mo>&times;</mo><msup><mn>10</mn><mn>%d</mn></msup></mstyle></math>", sign, sig, value, exponent10);break;
-	case 5: break;
-	case 6: fprintf(stdout, "%s%.*f",sign,sig,value);break;
-	default: break;
+        exp = exponent10%3;
+        idx = round(exponent10/3);
+        if( exponent10 > 0  ){
+            if(use_word == 0 ){ prefix = plus[idx]; } else { prefix = plus_word[idx]; }
+        }
+        else
+        {
+            if(use_word == 0){ prefix = min[-1*idx]; } else { prefix = min_word[-1*idx]; }
+        }
+        if( exp == 0){
+            fprintf(stdout, "%s%.*f %s",sign, sig, value,prefix);
+        }
+        else
+        {
+            fprintf(stdout, "%s%.*f&times;10<sup>%d</sup> %s", sign, sig, value, exp, prefix);
+        }
+        break;
+        case 4: fprintf(stdout, "<math xmlns=\"http://www.w3.org/1998/Math/MathML\" display=\"inline\"><mstyle id=\"wims_mathml\" mathsize=\"110%%\"><mn>%s%.*f</mn><mo>&times;</mo><msup><mn>10</mn><mn>%d</mn></msup></mstyle></math>", sign, sig, value, exponent10);break;
+        case 5: break;
+        case 6: fprintf(stdout, "%s%.*f",sign,sig,value);break;
+        default: break;
     }
     return NULL;
 }
@@ -213,8 +213,8 @@ char *printscience(double value, int sig, int format , int cnt ,int size){
 int main( int argc , char *argv[]){
 
     if( argc < 2){
-	fprintf(stdout,"syntax error : number1,significance1,type1 number2,significance2,type2 ... number_n,significance_n,type_n \n");
-	return 0;
+        fprintf(stdout,"syntax error : number1,significance1,type1 number2,significance2,type2 ... number_n,significance_n,type_n \n");
+        return 0;
     }
 
     double number = 0;
@@ -228,53 +228,53 @@ int main( int argc , char *argv[]){
     /* walk through argument 1 to end, and call function scienceprint(a,b,c) */
     input = argv[cnt];
     while( input != NULL ){
-	if(cnt > MAX_CONV){fprintf(stdout,"\nerror: number of conversions exceeds limit of %d\n",MAX_CONV);return 0;}
+        if(cnt > MAX_CONV){fprintf(stdout,"\nerror: number of conversions exceeds limit of %d\n",MAX_CONV);return 0;}
         while (*input){ /* loop through invalid chars. */
-	    if ( strchr(invalid_characters, *input) ){
-		fprintf(stdout,"\nerror : illegal character \"%s\" \n",input);
-		return 0;
-	    }
-	    input++;
-	}
-	/* reset input to actual value */
-	input = argv[cnt];
-	ptr = (char *) strtok(input,",");
-	idx = 0;
-	type = 0;
-	size = 0;
-	while( ptr != NULL ){
-	    switch( idx ){
-		case 0: 
-			/* size only interesting when 'significance=-1' 
-			 determine number of digits : 1.23445e+23 -> size = 6
-			*/
-			size = strlen(ptr);
-			if( strstr(ptr,".") != NULL){size = size - 1 ;}
-			if( strstr(ptr,"*10^") != NULL){
-			    ptr = str_replace(ptr,"*10^","E");
-			    if(ptr == NULL){
-				fprintf(stdout,"error : in replacement of 10^ notation\n");
-				return 0;
-			    } 
-			    size = size - 3;
-			}
-			if( strstr(ptr,"E") != NULL){size = size - strlen(strstr(ptr,"E"));}
-			if( strstr(ptr,"e") != NULL){size = size - strlen(strstr(ptr,"e"));}
-			number = atof(ptr); 
-			break;
-		case 1: significance = atoi(ptr);  break;
-		case 2: type = atoi(ptr); if(type < 0 || type > 5 ){type = 0;} break;
-		default: break;
-	    }
-	    idx++;
-	    ptr = (char *) strtok(NULL,",");
-	}
-	/* number and precision are mandatory:  default type=0  */
-	if( idx < 2 || idx > 3){fprintf(stdout,"\nsyntax error : number1,significance1,type1 number2,significance2,type2 ... number_n,significance_n,type_n \n");return 0;}
-	/* call conversion routine */
-	printscience(number, significance, type , cnt , size);
-	cnt++;
-	input = argv[cnt];
+            if ( strchr(invalid_characters, *input) ){
+                fprintf(stdout,"\nerror : illegal character \"%s\" \n",input);
+                return 0;
+            }
+            input++;
+        }
+        /* reset input to actual value */
+        input = argv[cnt];
+        ptr = (char *) strtok(input,",");
+        idx = 0;
+        type = 0;
+        size = 0;
+        while( ptr != NULL ){
+            switch( idx ){
+                case 0: 
+                        /* size only interesting when 'significance=-1' 
+                         determine number of digits : 1.23445e+23 -> size = 6
+                        */
+                        size = strlen(ptr);
+                        if( strstr(ptr,".") != NULL){size = size - 1 ;}
+                        if( strstr(ptr,"*10^") != NULL){
+                            ptr = str_replace(ptr,"*10^","E");
+                            if(ptr == NULL){
+                                fprintf(stdout,"error : in replacement of 10^ notation\n");
+                                return 0;
+                            } 
+                            size = size - 3;
+                        }
+                        if( strstr(ptr,"E") != NULL){size = size - strlen(strstr(ptr,"E"));}
+                        if( strstr(ptr,"e") != NULL){size = size - strlen(strstr(ptr,"e"));}
+                        number = atof(ptr); 
+                        break;
+                case 1: significance = atoi(ptr);  break;
+                case 2: type = atoi(ptr); if(type < 0 || type > 5 ){type = 0;} break;
+                default: break;
+            }
+            idx++;
+            ptr = (char *) strtok(NULL,",");
+        }
+        /* number and precision are mandatory:  default type=0  */
+        if( idx < 2 || idx > 3){fprintf(stdout,"\nsyntax error : number1,significance1,type1 number2,significance2,type2 ... number_n,significance_n,type_n \n");return 0;}
+        /* call conversion routine */
+        printscience(number, significance, type , cnt , size);
+        cnt++;
+        input = argv[cnt];
     }
     fprintf(stdout,"\n");
     return 0;
